weakref in scipy
